4 updated Document interface and refactored MainWindow implementation
- Document class accesses the Name and Description info from the Datastore

- MainWindow implementation broken up into several files for easier management
